+++ src/features/api/userApi.js
@@ -1,0 +1,12 @@
+import apiSlice from "./apiSlice";
+
+const userApi = apiSlice.injectEndpoints({
+    endpoints: builder => ({
+        postUserInfo: builder.mutation({
+            mutation: () => ({
+
+            })
+        })
+    })
+})
+
