+++ src/token/getToken.js
@@ -1,0 +1,16 @@
+export const getToken = (user) => {
+    const currentUser = {
+        email: user?.email
+    }
+    fetch(`http://localhost:5000/jwt`, {
+        method: 'POST',
+        headers: {
+            'content-type': 'application/json'
+        },
+        body: JSON.stringify(currentUser)
+    })
+        .then(res => res.json())
+        .then(data => {
+            localStorage.setItem('jwt-token', data.token);
+        })
+}
